requirements.txt.txt
streamlit
pandas
numpy
pillow
httpx
googletrans==4.0.0-rc1
plotly